Cleaning up approaches/ directory, deleting older implementations.
